llm-inference-server/requirements.txt
fastapi
uvicorn
mangum
mysql-connector-python
python-dotenv
pydantic
llama-index-llms-llama_api